% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_procs.R
\name{berkowitz_test}
\alias{berkowitz_test}
\title{Berkowitz-style test for calibration of multi step forecast distributions}
\usage{
berkowitz_test(u)
}
\arguments{
\item{u}{vector of PIT values}
}
\value{
p-value for the null hypothesis of calibration
}
\description{
Berkowitz-style test for calibration of multi step forecast distributions
}
\details{
Under the null of calibration, the unconditional distribution of u is standard uniform.
However, the elements of u may be serially correlated under the null. The present variant of the Berkowitz
test is described in Knueppel (2015). As described there, the test is not exact since the AR(1) model under the alternative
may be misspecified. However, the test often works well in practice.
}
\references{
Berkowitz, J. (2001): `Testing Density Forecasts, With Applications to Risk Management'.
Journal of Business and Economic Statistics 19, 465-474. \doi{10.1198/07350010152596718}

Knueppel, M. (2015): `Evaluating the Calibration of Multi-Step-Ahead Density Forecasts Using Raw Moments'.
Journal of Business and Economic Statistics 33, 270-281. \doi{10.1080/07350015.2014.948175}
}
