% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_procs.R
\name{score_calibration_es}
\alias{score_calibration_es}
\title{Computations for score calibration test}
\usage{
score_calibration_es(
  y,
  x,
  exact = FALSE,
  randomize = FALSE,
  use_C = TRUE,
  checks = TRUE
)
}
\arguments{
\item{y}{realization, vector of length d (= dimension of forecast distribution)}

\item{x}{forecast sample, matrix with d rows (variables) and m columns (forecast draws).}

\item{exact}{should exact estimator be used? If \code{FALSE} (the default), split sample in two halfs.}

\item{randomize}{should sample split be randomized? Defaults to \code{FALSE}. Only relevant if \code{exact = FALSE}.}

\item{use_C}{should Cpp implementation be used? Defaults to \code{TRUE}; alternatively, use base R code based on \code{stats::dist}. Should yield the same results. C code is usually faster.}

\item{checks}{should inputs be checked for consistency? Defaults to \code{TRUE}.}
}
\value{
list with three components, \code{u}, \code{d} and \code{es}. All are scalar numbers. \code{u} is the PIT value of the realized score within the simulated scores. \code{d} is the difference between the realized score and the expected score. \code{es} is the energy score.
}
\description{
Computations for score calibration test
}
\examples{
set.seed(1)
# Draw five-dimensional forecast distribution and realization
x <- matrix(rnorm(5000), 5, 1000)
y <- rnorm(5)
# Compute score calibration quantities
score_calibration_es(y, x)
# Compare to scoringRules function (should yield the same
# energy score value if exact version is used)
scoringRules::es_sample(y, x)
}
