% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_procs.R
\name{t_hac}
\alias{t_hac}
\title{Compute autocorrelation-robust t statistic}
\usage{
t_hac(x, implementation = "MK", variance_under_h0 = FALSE, ...)
}
\arguments{
\item{x}{vector of data (with mean = 0 under null hypothesis)}

\item{implementation}{string, either "MK" or "sandwich". The former is the implementation of Knueppel (2015), the
latter uses the function \code{NeweyWest} from the sandwich package}

\item{variance_under_h0}{logical, should null hypothesis be imposed for variance estimation?}

\item{...}{additional parameters that can be passed to the robust variance function}
}
\description{
Compute autocorrelation-robust t statistic
}
\examples{
# Simulate AR(1) model with zero mean
y <- arima.sim(model = list(ar = .5), n = 500)
# Different test variants
t_hac(y, implementation = 'MK')
t_hac(y, implementation = 'MK', variance_under_h0 = TRUE)
t_hac(y, implementation = 'sandwich')
}
